Add vectored exception handling sample. Change hierarchy to output built files for cpp projects
